+++ Locales/enUS.lua
@@ -52,6 +52,15 @@ select(2, ...).L = setmetatable({
     ["RAID_DEBUFFS_TIPS"] = "Tips: [Drag & Drop] to change debuff order. [Double-Click] on instance name to open Encounter Journal. [Shift+Left Click] on instance/boss name to share debuffs. [Alt+Left Click] on instance/boss name to reset debuffs. The priority of General Debuffs is higher than Boss Debuffs.",
 
     ["CHANGE LOGS"] = [[
+        <h1>r93-release (Apr 16, 2022, 06:45 GMT+8)</h1>
+        <p>+ Added an indicator: Externals + Defensives.</p>
+        <p>+ Added a new custom indicator type: texture.</p>
+        <p>+ Implemented import &amp; export for all settings (check About tab).</p>
+        <p>+ Implemented layout auto switch for Mythic (raid).</p>
+        <p>* Updated zhTW.</p>
+        <p>* Fixed some bugs.</p>
+        <br/>
+
         <h1>r92-release (Apr 12, 2022, 14:30 GMT+8)</h1>
         <p>* Fixed health color (gradient).</p>
         <br/>