+++ CHANGELOG.md
@@ -1,4 +1,11 @@
-[Full Changelog](https://github.com/enderneko/Cell/compare/r91-release...6a008de8f95bf6b357991c5f111037dc42f877a3)
+[Full Changelog](https://github.com/enderneko/Cell/compare/r92-release...5763aee4a92d0698d56d8fe4fea174c7de437219)
 
-- Fix marks texture
-- Fix gradient health color
+- Add Externals+Defensives indicator
+- Add custom indicator type Texture
+- Implement All Settings import & export
+- Add layout auto switch for Mythic
+- Replace my ColorPicker with original
+- Force update boss678 more often
+- Bump unsupported version to r81
+- Update option frame sizes
+- Update zhTW